% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.alpha.R
\name{f.alpha}
\alias{f.alpha}
\title{Incremental Alpha for Group Sequential Design}
\usage{
f.alpha(overall.alpha = 0.025, sf = "LDOF", timing = c(0.75, 1), param = -3)
}
\arguments{
\item{overall.alpha}{Allocated overall alpha (one-sided) for group sequential design}

\item{sf}{Spending function. LanDeMets O'Brien Fleming: "LDOF", LanDeMets Pocock: "LDPK", "HSD": Hwang-Shih-DeCani spending function with parameter param.}

\item{timing}{Information fractions of analyses, for example timing = c(0.6, 0.75, 1.0).}

\item{param}{parameter for Hwang-Shih-DeCani spending function}
}
\description{
The incremental alpha is the type I error allocated to a specific analysis while the sequential tests are negative for all previous analyses. Only one-sided type I error (alpha) is considered.
Specifically, incremental alpha_k = P(reject H0 at kth analysis while not rejecting H0 at all previous analyses)
}
\examples{
(1) Group sequential tests at information times 0.5, 0.7, 1.0 with overall alpha 0.025.

#LD OBF spending
f.alpha(overall.alpha=0.025, sf="LDOF", timing=c(0.5, 0.7, 1.0))

#LD Pocock spending
f.alpha(overall.alpha=0.025, sf="LDPK", timing=c(0.5, 0.7, 1.0))

#HSD(-3)
f.alpha(overall.alpha=0.025, sf="HSD", timing=c(0.5, 0.7, 1.0), param=-3)


}
