% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.dataCut.R
\name{f.dataCut}
\alias{f.dataCut}
\title{Data Cut Utility Function to Determine the Datasets According to Target Events}
\usage{
f.dataCut(data, targetEvents = 397, DCO = NULL)
}
\arguments{
\item{data}{Input dataset}

\item{targetEvents}{A vector of targetEvents for planned analyses}

\item{DCO}{A vector of data cutoffs for planned analyses. Either targetEvents or DCO is required.}
}
\value{
An object with the components:
\describe{
\item{[\link{1}]}{Dataset for First Analysis}
\item{[\link{2}]}{Dataset for Second Analysis}
\item{[\link{L}]}{Dataset for Lth Analysis}
}
}
\description{
Determine the dataset for each analysis according to specified target event.
}
