% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.power.HaybittlePeto.R
\name{lr.power.HaybittlePeto}
\alias{lr.power.HaybittlePeto}
\title{Power Calculation Using Haybittle-Peto Boundary and Logrank Test in Group Sequential Design}
\usage{
lr.power.HaybittlePeto(
  p1 = 3e-04,
  overall.alpha = 0.025,
  events = c(96, 153, 250),
  events0 = NULL,
  events1 = NULL,
  hr = 0.6,
  r = 1,
  variance = "H0"
)
}
\arguments{
\item{p1}{A fixed p value boundary for IAs}

\item{overall.alpha}{Overall alpha for the test in the group sequential design.
Default, overall.alpha = 0.025. overall.alpha must be one-sided.}

\item{events}{A vector of target events for IAs and FA}

\item{events0}{A vector of target events for IAs and FA for control arm.
Required when variance option is H1.}

\item{events1}{A vector of target events for IAs and FA for experimental arm
Required when variance option is H1.}

\item{hr}{Hazard ratio (smaller better)}

\item{r}{Randomization ratio. r=1 means 1:1; r=2 means 2:1.}

\item{variance}{Option for variance estimate. "H1" or "H0". Default H1, which is usually more conservative than H0.}
}
\value{
Objects of two dataframes alpha and design
\itemize{
\item alpha      Incremental alpha
\item cum.alpha  Cumulative alpha
\item overall.alpha    Overall alpha
\item side       Always one-sided
\item events     Number of events by each analysis
\item information Information fraction by each analysis
\item p          Rejection boundary in p value by each analysis
\item z          Rejection boundary in z value by each analysis
\item cv         Critical value (minimum detectable difference) in hazard ratio
\item marg.power Marginal power, defined as the power at each analysis assuming as a separate study. Note: This is for reference only and it is not consistent with group sequential design.
\item incr.power Incremental power by each analysis
\item cum.power Cumulative power by each analysis
\item overall.power Overall Power
}
}
\description{
This function calculates the power for the given number of events and rejection boundary
at each analysis for group sequential design
}
\examples{

#(1) 
lr.power.HaybittlePeto(p1=0.0003, overall.alpha=0.025, hr = 0.7, r = 1, events=c(96, 153, 250))

#(1) 
lr.power.HaybittlePeto(p1=0.0003, overall.alpha=0.025, hr = 0.7, r = 1, events=c(229.4766, 340.8875))

#(2) O'Brien Fleming alpha spending, variance under H1
lr.power.HaybittlePeto(p1=0.0003, overall.alpha=0.025, hr = 14/22, r = 1, events=c(210, 280))
                     
}
